+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -19,16 +19,12 @@ import {
   loadSearchConfigurationActions,
   SearchEngineConfiguration,
 } from '@coveo/headless';
-import {
-  AtomicStore,
-  Bindings,
-  InitializeEvent,
-  initialStore,
-} from '../../utils/initialization-utils';
+import {Bindings, InitializeEvent} from '../../utils/initialization-utils';
 import i18next, {i18n} from 'i18next';
 import Backend, {BackendOptions} from 'i18next-http-backend';
 import {createStore} from '@stencil/store';
 import {setCoveoGlobal} from '../../global/environment';
+import {AtomicStore, initialStore} from '../../utils/store';
 
 export type InitializationOptions = Pick<
   SearchEngineConfiguration,
@@ -50,7 +46,7 @@ export class AtomicSearchInterface {
   private initialized = false;
   private store = createStore<AtomicStore>(initialStore());
 
-  @Element() private host!: HTMLDivElement;
+  @Element() private host!: HTMLElement;
 
   @State() private error?: Error;
 
@@ -255,7 +251,12 @@ export class AtomicSearchInterface {
   }
 
   private get bindings(): Bindings {
-    return {engine: this.engine!, i18n: this.i18n, store: this.store};
+    return {
+      engine: this.engine!,
+      i18n: this.i18n,
+      store: this.store,
+      interfaceElement: this.host,
+    };
   }
 
   private initComponents() {
@@ -309,7 +310,7 @@ export class AtomicSearchInterface {
     return [
       this.engine && (
         <atomic-relevance-inspector
-          bindings={{engine: this.engine, i18n: this.i18n, store: this.store}}
+          bindings={this.bindings}
         ></atomic-relevance-inspector>
       ),
       <slot></slot>,